- book/makebook2021.cpp 追加。これから書く。 - book.cppにあったmakebookコマンドの処理ルーチン、book/makebook2015.cppに移動。 	- 2015年ごろに書いたコードだから。 - Aperyの定跡読み込みで使っているMT64bit、apery_book.cppでしか使っていないので無名namespaceに変更。
